requirements.txt
streamlit
google-genai
openpyxl